requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pyyaml>=6.0.1
python-dateutil>=2.8.2
rich>=13.0.0
